+++ packages/nhost-js/CHANGELOG.md
@@ -1,4 +1,5 @@
 # dev
 
+- fix (tools/codegen;packages/nhost-js): encode URL parameters better #30
 - fix (packages/next-js): use proper types for Webauthn endpoints #29
 - chore (ci): enabled #26